feat: Customize the button to exit full screen mode
